# server/src/swagger/routes.yaml
swagger: '3.0.0'
info:
  title: Kids Activity Tracker API
  version: '1.0.0'
  description: |
    Comprehensive API documentation for the Kids Activity Tracker application.
    
    ## API Overview
    
    ### Core Concepts:
    - **Categories**: 5 age-based groupings (Early Years: Parent Participation, Early Years: On My Own, School Age, Youth, All Ages & Family)
    - **Activity Types**: 22 subject-based groupings (Swimming & Aquatics, Team Sports, etc.)
    - **Global Filters**: hideClosedActivities, hideFullActivities (apply to all endpoints except Favorites)
    - **Pagination**: 50 items per page with offset-based pagination
    
    ### Navigation Flows:
    1. **Browse by Location**: Dashboard → Cities → Venues → Activities
    2. **Browse by Category**: Dashboard → Categories → Category Detail → Activities  
    3. **Browse by Activity Type**: Dashboard → Activity Types → Activity Type Detail → Activities
    4. **Recommendations**: Dashboard → Recommended Activities (user-specific)
    
    ### React Native Components Usage:
    - **DashboardScreen**: Uses categories, activity types, locations APIs
    - **CategoryDetailScreen**: Uses `/categories/{id}/activities`
    - **AllCategoriesScreen**: Uses `/categories`
    - **LocationBrowseScreen**: Uses `/locations/cities`, `/locations/{cityId}/venues`
    - **ActivityDetailScreen**: Uses `/activities/{id}`, favorites APIs
    - **SearchScreen**: Uses `/activities/search` with extensive filtering
paths:
...
  /api/v1/activities:
    get:
      tags: [Activities]
      summary: Search and filter activities
      description: |
        Primary activity search endpoint with comprehensive filtering.
        Used by: SearchScreen, FilterScreen, RecommendedActivitiesScreen
        Supports all global filters and advanced search capabilities.
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search in activity name, description
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of activities to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of activities to skip
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
          description: Filter out closed activities
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
          description: Filter out full activities
        - in: query
          name: activityTypes
          schema:
            type: array
            items:
              type: string
          description: Filter by activity type IDs or names
        - in: query
          name: categories
          schema:
            type: array
            items:
              type: string
          description: Filter by category IDs
        - in: query
          name: locations
          schema:
            type: array
            items:
              type: string
          description: Filter by location IDs or city names
        - in: query
          name: ageMin
          schema:
            type: integer
          description: Minimum age filter
        - in: query
          name: ageMax
          schema:
            type: integer
          description: Maximum age filter
        - in: query
          name: costMin
          schema:
            type: number
          description: Minimum cost filter
        - in: query
          name: costMax
          schema:
            type: number
          description: Maximum cost filter
      responses:
        200:
          description: Paginated search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
                  filters:
                    type: object
                    description: Applied filters summary
  /api/v1/activities/{id}:
    get:
      tags: [Activities]
      summary: Get activity details
      description: |
        Get detailed information about a specific activity.
        Used by: ActivityDetailScreen, ActivityCard navigation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Activity UUID
      responses:
        200:
          description: Activity details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activity:
                    $ref: '#/components/schemas/Activity'
        404:
          description: Activity not found
...
  /api/favorites:
    get:
      tags: [Favorites]
      summary: Get user's favorite activities
      description: |
        Returns paginated list of user's favorite activities.
        Used by: FavoritesScreen, FavoritesStack
        NOTE: This endpoint does NOT apply global filters (hideClosedActivities, etc.)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: User's favorite activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  favorites:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        activity:
                          $ref: '#/components/schemas/Activity'
                        createdAt:
                          type: string
                          format: date-time
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
    
    post:
      tags: [Favorites]
      summary: Add activity to favorites
      description: |
        Add an activity to user's favorites.
        Used by: ActivityCard, ActivityDetailScreen favorite button
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                activityId:
                  type: string
                  format: uuid
              required: [activityId]
      responses:
        201:
          description: Added to favorites
        409:
          description: Already in favorites
  /api/favorites/{id}:
    delete:
      tags: [Favorites]
      summary: Remove activity from favorites
      description: |
        Remove an activity from user's favorites.
        Used by: FavoritesScreen, ActivityDetailScreen unfavorite action
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Favorite record UUID (not activity ID)
      responses:
        200:
          description: Removed from favorites
        404:
          description: Favorite not found

  # ==================== CHILD ACTIVITIES ENDPOINTS ====================
...
  /api/v1/locations/cities:
    get:
      tags: [Locations]
      summary: Get all cities with venue and activity counts
      description: |
        Returns all cities with their venue and activity counts.
        Used by: LocationBrowseScreen for city selection.
        Navigation flow: Dashboard → Browse by Location → Cities (this endpoint) → Venues → Activities
      parameters:
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
          description: Filter out closed activities from counts
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
          description: Filter out full activities from counts
        - in: query
          name: includeActivityCounts
          schema:
            type: boolean
            default: true
          description: Whether to include activity counts (can be disabled for performance)
      responses:
        200:
          description: List of cities with counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cities:
                    type: array
                    items:
                      $ref: '#/components/schemas/City'
...
  /api/v1/categories:
    get:
      tags: [Categories]
      summary: Get all age-based categories with activity counts
      description: |
        Returns all 5 age-based categories (Early Years: Parent Participation, Early Years: On My Own, 
        School Age, Youth, All Ages & Family) with activity counts.
        Used by: DashboardScreen "Browse by Category" section, AllCategoriesScreen.
        Navigation flow: Dashboard → Categories → Category Detail → Activities
      parameters:
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
          description: Filter out closed activities from counts
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
          description: Filter out full activities from counts
        - in: query
          name: includeActivityCounts
          schema:
            type: boolean
            default: true
          description: Whether to include activity counts
      responses:
        200:
          description: List of age-based categories with counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
...
components:
  schemas:
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        cost:
          type: number
        spotsAvailable:
          type: integer
        totalSpots:
          type: integer
        registrationStatus:
          type: string
          enum: [Open, Closed, Full, Cancelled]
        dateStart:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        registrationDate:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        activityType:
          $ref: '#/components/schemas/ActivityType'
        provider:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            website:
              type: string
    ActivityType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        activityCount:
          type: integer
        subtypes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              code:
                type: string
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          enum: [
            "Early Years: Parent Participation",
            "Early Years: On My Own", 
            "School Age",
            "Youth",
            "All Ages & Family"
          ]
        description:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        requiresParent:
          type: boolean
        activityCount:
          type: integer
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        city:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            province:
              type: string
            country:
              type: string
        activityCount:
          type: integer
    City:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        province:
          type: string
        country:
          type: string
        venueCount:
          type: integer
        activityCount:
          type: integer
...
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        pages:
          type: integer
          description: Total number of pages
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...